+++ lib/unicode/script.ex
@@ -11,8 +11,36 @@ defmodule Unicode.Script do
   end
 
   @known_scripts Map.keys(@scripts)
+
   def known_scripts do
     @known_scripts
+  end
+
+  @script_alias Utils.property_value_alias()
+  |> Map.get("sc")
+  |> Enum.flat_map(fn
+      [alias1, code] ->
+        [{String.downcase(alias1), String.downcase(code)}]
+      [alias1, code, alias2] ->
+        [{String.downcase(alias1), String.downcase(code)},
+         {String.downcase(alias2), String.downcase(code)}]
+  end)
+  |> Map.new
+
+  def aliases do
+    @script_alias
+  end
+
+  def fetch(script) do
+    script = Map.get(aliases(), script, script)
+    Map.fetch(scripts(), script)
+  end
+
+  def get(script) do
+    case fetch(script) do
+      {:ok, script} -> script
+      _ -> nil
+    end
   end
 
   @doc """
@@ -26,9 +54,9 @@ defmodule Unicode.Script do
 
   """
   def count(script) do
-    scripts()
-    |> Map.get(script)
-    |> Enum.reduce(0, fn {from, to}, acc -> acc + to - from + 1 end)
+    with {:ok, script} <- fetch(script) do
+      Enum.reduce(script, 0, fn {from, to}, acc -> acc + to - from + 1 end)
+    end
   end
 
   def script(string) when is_binary(string) do